moves embed into module
