+++ package.json
@@ -11,8 +11,8 @@
     "elm-build": "elm make src/elm/Main.elm --output=public/main.js --optimize && npm run compress-main",
     "css-dev": "postcss --watch --config ./postcss-config.js",
     "css-build": "postcss --config ./postcss-config.js",
-    "js-dev": "watchify src/js/embed.js -o public/embed.js -v",
-    "js-build": "browserify src/js/embed.js -o public/embed.js -t [ babelify --presets [ env ] ]",
+    "js-dev": "watchify src/js/main.js -o public/embed.js -v",
+    "js-build": "browserify src/js/main.js -o public/embed.js -t [ babelify --presets [ env ] ]",
     "copy-pep": "cp node_modules/pepjs/dist/pep.min.js public",
     "sw-generate": "workbox generateSW",
     "compress-main": "uglifyjs public/main.js --compress 'pure_funcs=\"F2,F3,F4,F5,F6,F7,F8,F9,A2,A3,A4,A5,A6,A7,A8,A9\",pure_getters=true,keep_fargs=false,unsafe_comps=true,unsafe=true,passes=2' --output=public/main.js && uglifyjs public/main.js --mangle --output=public/main.js",
