+++ src/js/cheats.js
@@ -1,0 +1,12 @@
+const { setProgress } = require("./cache.js");
+
+function activate() {
+  window.skipToLevel = skipToLevel;
+}
+
+function skipToLevel(worldId, levelId) {
+  setProgress({ worldId, levelId });
+  window.location.reload();
+}
+
+module.exports = { activate };
