+++ src/js/main.js
@@ -1,0 +1,6 @@
+const { Embed } = require("./embed.js");
+const cheats = require("./cheats.js");
+
+cheats.activate();
+
+Embed(Elm.Main);
