+++ src/js/serviceworker.js
@@ -1,0 +1,9 @@
+function register() {
+  if ("serviceWorker" in navigator) {
+    window.addEventListener("load", () => {
+      navigator.serviceWorker.register("./sw.js");
+    });
+  }
+}
+
+module.exports = { register };
